Ocenjivanje preparata od strane korisnika

+++ backend/service/src/main/java/com/ftn/sbnz/respository/PreparatRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.ftn.sbnz.model.Preparat;
 
 @Repository
-public interface PreparatRepository extends JpaRepository<Preparat, Integer>{ 
+public interface PreparatRepository extends JpaRepository<Preparat, Long>{ 
     
     // @Query("SELECT p FROM Preparat inner join preparat_ocena using(preparat_id) gorup by preparat_id order by avg(ocene) desc")
     // List<Preparat> getAllSortedByRating();
